Refactor custom errors
