--- services/nft-image-indexer/customErrors/unsupported-svg-error.go
@@ -1,23 +1,0 @@
-package errors
-
-import (
-	"errors"
-	"fmt"
-)
-
-type UnsupportedSVG struct {
-	Url string
-
-	Err error
-}
-
-func (r *UnsupportedSVG) Error() string {
-	return fmt.Sprintf("unsupported SVG Url: %v", r.Url)
-}
-
-func NewUnsupportedSVG(url string) error {
-	return &UnsupportedSVG{
-		Url: url,
-		Err: errors.New("unsupported SVG URL"),
-	}
-}
