+++ services/nft-image-indexer/imageStore/errors.go
@@ -1,7 +1,8 @@
-package errors
+package imageStore
 
 import (
 	"errors"
+	"fmt"
 )
 
 const (
@@ -31,5 +32,22 @@ func NewImageCachingError(name string) error {
 	return &ImageCachingError{
 		Name: name,
 		Err:  ImageCachingErrorTypes[name],
+	}
+}
+
+type UnsupportedSVG struct {
+	Url string
+
+	Err error
+}
+
+func (r *UnsupportedSVG) Error() string {
+	return fmt.Sprintf("unsupported SVG Url: %v", r.Url)
+}
+
+func NewUnsupportedSVG(url string) error {
+	return &UnsupportedSVG{
+		Url: url,
+		Err: errors.New("unsupported SVG URL"),
 	}
 }
