+++ services/nft-image-indexer/imageStore/store.go
@@ -7,7 +7,6 @@ import (
 	"strings"
 	"time"
 
-	Errors "github.com/bitmark-inc/nft-indexer/services/nft-image-indexer/customErrors"
 	"github.com/cloudflare/cloudflare-go"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
@@ -117,14 +116,14 @@ func (s *ImageStore) UploadImage(ctx context.Context, assetID string, imageDownl
 		downloadStartTime := time.Now()
 		file, mimeType, err := imageDownloader.Download()
 		if err != nil {
-			return Errors.NewImageCachingError(Errors.ErrDownloadFileError)
+			return NewImageCachingError(ErrDownloadFileError)
 		}
 		logrus.
 			WithField("duration", time.Since(downloadStartTime)).
 			WithField("assetID", assetID).Debug("download thumbnail finished")
 
 		if !IsSupportedImageType(mimeType) {
-			return Errors.NewImageCachingError(Errors.ErrUnsupportImageType)
+			return NewImageCachingError(ErrUnsupportImageType)
 		}
 
 		if metadata == nil {
@@ -149,7 +148,7 @@ func (s *ImageStore) UploadImage(ctx context.Context, assetID string, imageDownl
 		if err != nil {
 			isErrSizeTooLarge, _ := regexp.MatchString("entity.*too large", err.Error())
 			if isErrSizeTooLarge {
-				return Errors.NewImageCachingError(Errors.ErrSizeTooLarge)
+				return NewImageCachingError(ErrSizeTooLarge)
 			}
 			return err
 		}
